ai-native-dev-main 2/practices/practice-03-chatapp/backend/pyproject.toml
[project]
name = "openai-llm-app-to-rest-api"
version = "0.1.0"
description = "Simple FastAPI with OpenAI integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.2",
    "openai>=2.3.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
